Move cron functions master.go -> monitor.go
